+++ Robust.Shared/Physics/Dynamics/PhysicsMapComponent.cs
@@ -21,8 +21,8 @@
 */
 
 using System.Collections.Generic;
-using System.Linq;
 using Robust.Shared.GameObjects;
+using Robust.Shared.GameStates;
 using Robust.Shared.IoC;
 using Robust.Shared.Log;
 using Robust.Shared.Map;
@@ -34,7 +34,8 @@ using PhysicsComponent = Robust.Shared.Physics.Components.PhysicsComponent;
 
 namespace Robust.Shared.Physics.Dynamics
 {
-    public abstract class SharedPhysicsMapComponent : Component
+    [RegisterComponent, NetworkedComponent]
+    public sealed class PhysicsMapComponent : Component
     {
         [Dependency] private readonly IEntityManager _entityManager = default!;
 
@@ -139,10 +140,10 @@ namespace Robust.Shared.Physics.Dynamics
     public readonly struct PhysicsUpdateBeforeMapSolveEvent
     {
         public readonly bool Prediction;
-        public readonly SharedPhysicsMapComponent MapComponent;
+        public readonly PhysicsMapComponent MapComponent;
         public readonly float DeltaTime;
 
-        public PhysicsUpdateBeforeMapSolveEvent(bool prediction, SharedPhysicsMapComponent mapComponent, float deltaTime)
+        public PhysicsUpdateBeforeMapSolveEvent(bool prediction, PhysicsMapComponent mapComponent, float deltaTime)
         {
             Prediction = prediction;
             MapComponent = mapComponent;
@@ -154,10 +155,10 @@ namespace Robust.Shared.Physics.Dynamics
     public readonly struct PhysicsUpdateAfterMapSolveEvent
     {
         public readonly bool Prediction;
-        public readonly SharedPhysicsMapComponent MapComponent;
+        public readonly PhysicsMapComponent MapComponent;
         public readonly float DeltaTime;
 
-        public PhysicsUpdateAfterMapSolveEvent(bool prediction, SharedPhysicsMapComponent mapComponent, float deltaTime)
+        public PhysicsUpdateAfterMapSolveEvent(bool prediction, PhysicsMapComponent mapComponent, float deltaTime)
         {
             Prediction = prediction;
             MapComponent = mapComponent;